+++ src/models.rs
@@ -14,16 +14,16 @@ use utils::syslog;
 #[derive(Serialize)]
 pub struct Sensors {
     pub label: String,
-    pub temp: f32,
+    pub temp: f64,
 }
 
 #[derive(Serialize)]
 pub struct Data {
     pub os: String,
     pub hostname: String,
-    pub uptime: u64,
+    pub uptime: i64,
     pub uuid: String,
-    pub cpu_freq: u64,
+    pub cpu_freq: i64,
     pub sensors: Vec<Sensors>,
     pub user: String,
     pub mac_address: String,
